Added information panel and fixed stop button
The stop button was only acting as a pause button, and did not reset the playback to the first frame. It now resets the video to the first frame, and displays the first frame to the user.

An information panel has been added to display the current frame number in the scene. This may be useful for debugging.

There remains a problem where in order to render the first frame, the update() function is called, which "steps" the frame to the second frame, in order to render an initial frame.